Refactor WelcomeBar component by updating import path for improved organization
