+++ src/components/groups/welcome-bar/WelcomeBar.tsx
